refactor/model-ingest (#148)
* moved non-handler functions to separate script inside calcloud

* add missing source path for lambda itself

* black and flake8
