+++ terraform/lambda_model_ingest.tf
@@ -1,8 +1,6 @@
 resource "aws_dynamodb_table" "calcloud_model_db" {
   name           = "calcloud-model${local.environment}"
   billing_mode   = "PAY_PER_REQUEST" #"PROVISIONED"
-  #read_capacity  = 10
-  #write_capacity = 10
   hash_key       = "ipst"
 
   attribute {
@@ -31,9 +29,16 @@ module "lambda_model_ingest" {
 
   source_path = [
     {
+      # this is the lambda itself. The code in path will be placed directly into the lambda execution path
       path = "${path.module}/../lambda/ModelIngest"
       pip_requirements = true
     },
+    {
+      # calcloud for the package
+      path = "${path.module}/../calcloud"
+      prefix_in_zip = "calcloud"
+      pip_requirements = false
+    }
   ]
 
   store_on_s3 = true
